pyproject.toml
[project]
name = "thissentencedoesnotexist"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "carterwward", email = "Carter.Ward@uvm.edu"},
]
dependencies = ["litestar>=2.21.1", "psycopg2-binary>=2.9.11", "pydantic>=2.12.5"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
